requirements.txt
google-generativeai
pillow 
python-dotenv
requests
streamlit